Move ConfigureConnectorSettings to Domain project
